move test models and factories to app
