+++ tests/App/Models/Post.php
@@ -1,25 +1,37 @@
 <?php
 
-namespace Abbasudo\Purity\Tests\Models;
+namespace Abbasudo\Purity\Tests\App\Models;
 
 use Abbasudo\Purity\Traits\Filterable;
 use Abbasudo\Purity\Traits\Sortable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
-class Project extends Model
+class Post extends Model
 {
     use HasFactory;
     use Filterable;
     use Sortable;
 
     protected $fillable = [
-        'name',
+        'title',
     ];
+
+    public function comments(): HasMany
+    {
+        return $this->hasMany(Comment::class);
+    }
 
     public function user(): BelongsTo
     {
         return $this->belongsTo(User::class);
+    }
+
+    public function tags(): BelongsToMany
+    {
+        return $this->belongsToMany(Tag::class);
     }
 }
