+++ tests/App/Models/Project.php
@@ -1,28 +1,25 @@
 <?php
 
-namespace Abbasudo\Purity\Tests\Models;
+namespace Abbasudo\Purity\Tests\App\Models;
 
-use Abbasudo\Purity\Tests\Factories\ProductFactory;
 use Abbasudo\Purity\Traits\Filterable;
 use Abbasudo\Purity\Traits\Sortable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
-class Product extends Model
+class Project extends Model
 {
     use HasFactory;
     use Filterable;
     use Sortable;
 
-    protected static function newFactory()
-    {
-        return ProductFactory::new();
-    }
-
     protected $fillable = [
         'name',
-        'price',
-        'rate',
-        'is_available',
     ];
+
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class);
+    }
 }
