+++ tests/App/Models/Tag.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Abbasudo\Purity\Tests\Models;
+namespace Abbasudo\Purity\Tests\App\Models;
 
-use Abbasudo\Purity\Tests\Filters\CustomFilterResolver;
+use Abbasudo\Purity\Tests\App\Filters\CustomFilterResolver;
 use Abbasudo\Purity\Traits\Filterable;
 use Abbasudo\Purity\Traits\Sortable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
