+++ tests/Feature/RestrictFilterableFieldsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use Abbasudo\Purity\Tests\Models\Post;
+use Abbasudo\Purity\Tests\App\Models\Post;
 use Abbasudo\Purity\Tests\TestCase;
 use Illuminate\Support\Facades\Route;
